Optimize tests
