+++ tests/file_fixtures/docker.txt
@@ -1,0 +1,2 @@
+FROM python:3.9-buster as production
+ARG POETRY_VERSION=1.1.4
