+++ tests/file_fixtures/pyproject.txt
@@ -1,0 +1,10 @@
+[tool.poetry]
+name = 'test'
+version = '0.1.0'
+description = 'test'
+authors = ['test']
+license = 'MIT'
+
+[tool.poetry.dependencies]
+python = '^3.9'
+requests = '^2.31.0'
