+++ tests/test_repository.py
@@ -1,34 +1,4 @@
-import os
-import pathlib
 import subprocess
-
-import pytest
-from src.managers.repository import RepositoryManager
-
-
-@pytest.fixture
-def repo(tmpdir):
-    t = tmpdir.mkdir("repo")
-    pyproject_content = "[tool.poetry]\nname = 'test'\nversion = '0.1.0'\ndescription = 'test'\nauthors = ['test']\nlicense = 'MIT'\n[tool.poetry.dependencies]\npython = '^3.9'\nrequests = '^2.31.0'\n"  # noqa: E501
-    t.join("pyproject.toml").write(pyproject_content)
-    dockerfile_content = "FROM python:3.9-buster as production\nARG POETRY_VERSION=1.1.4\n"
-    t.join("Dockerfile").write(dockerfile_content)
-
-    os.chdir(t)
-    subprocess.run(["git", "config", "--global", "user.email", "user@example.com"], check=True, capture_output=True)
-    subprocess.run(["git", "config", "--global", "user.name", "user"], check=True, capture_output=True)
-    subprocess.run(["git", "init"], check=True, capture_output=True)
-    subprocess.run(["git", "add", "."], check=True, capture_output=True)
-    subprocess.run(["git", "commit", "-m", "initial commit"], check=True, capture_output=True)
-    subprocess.run(["git", "checkout", "master"], check=True, capture_output=True)
-    subprocess.run(["git", "checkout", "-b", "test"], check=True, capture_output=True)
-    path = pathlib.Path(t)
-    return path
-
-
-@pytest.fixture
-def repository_manager(repo):
-    return RepositoryManager(repo_url=repo, docker_file_name="Dockerfile")
 
 
 def test_repository_manager_init(repository_manager, repo):
